JavaScriptファイルの分割、Bootstrap Iconsのインポート削除 #35
